Module_4_Analytics_Engineering/requirements.txt
google-cloud-storage==3.0.0
google-cloud-bigquery==3.27.0
dbt-bigquery==1.9.1
requests==2.32.3
